optimize: collapse four real duplications (agent 1)
- lib/open-browser.ts is now the one home for the platform browser-open,
  previously byte-identical in api/grok-auth and lib/mcp-oauth-server
- api/mcp PUT calls saveFile() instead of re-implementing its
  read-sha/update-or-create/commit sequence
- CLI: requireGh() in output.ts replaces the gh-auth guard copy-pasted at
  8 sites (secrets.ts:32 keeps its distinct message)
- api-client.ts gains getJson<T>(); page.tsx's 6 hand-rolled fetch-and-parse
  expressions use it, with every existing .catch/.then handler preserved

Left duplicated on purpose: the memory routes, the two build routes,
buildSoul/buildStrategy, and the CLI subcommand preamble.
