+++ apps/cli/src/commands/telegram.ts
@@ -1,5 +1,5 @@
-import { dispatchCommandsWorkflow, ghAvailable } from '../../../dashboard/lib/gh.ts'
-import { emit, c, fail, isDryRun } from '../output.ts'
+import { dispatchCommandsWorkflow } from '../../../dashboard/lib/gh.ts'
+import { emit, c, fail, isDryRun, requireGh } from '../output.ts'
 
 const USAGE = `aeon telegram — Telegram bot integration
 
@@ -20,7 +20,7 @@ export function telegramCommand(argv: string[]) {
   if (isDryRun()) return emit({ dryRun: true, workflow: 'setup-commands.yml' }, () =>
     console.log(c.yellow('dry-run: ') + 'gh workflow run setup-commands.yml'))
 
-  if (!ghAvailable()) fail('GitHub CLI not authenticated. Run: gh auth login')
+  requireGh()
   try { dispatchCommandsWorkflow() } catch (e) { fail(e instanceof Error ? e.message : 'failed to dispatch') }
   emit({ ok: true }, () => console.log(c.green('✓ ') + 'dispatched setup-commands.yml — the / menu will refresh shortly'))
 }
